Software/dashboard-backend-main/dashboard-backend-main/requirements.txt
flask==3.0.3
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Flask-Cors==5.0.0
Flask-SocketIO==5.5.0
SQLAlchemy-Utils==0.41.2